[test] Move stl-cmath's test_func helpers into TestUtils. NFC
test/Features/stl-cmath.cpp defined get_tolerance, get_h, numerical_derivative
and test_func locally. Move them to test/TestUtils.h so there is a single
definition of the forward-vs-reverse-vs-numerical check, and let stl-cmath.cpp
use them from there.

TestUtils.h now includes Differentiator.h: test_func names clad::CladFunction
and clad::gradient in its signature, so the header has to stay self-contained.

No test is added and no behaviour changes; stl-cmath.cpp's macros keep the same
expectations and output.
