+++ test/Features/stl-cmath.cpp
@@ -131,82 +131,37 @@ namespace std {
   }
 }
 #endif
-
-template <typename T>
-T get_tolerance() {
-    if constexpr (std::is_same_v<T,float>) return T(1e-3);
-    else if constexpr (std::is_same_v<T,long double>) return T(1e-10);
-    else return T(1e-6);
-}
-
-// step for numerical derivative
-template <typename T>
-T get_h() {
-    if constexpr (std::is_same_v<T,float>) return T(1e-3);
-    else if constexpr (std::is_same_v<T,long double>) return T(1e-8);
-    else return T(1e-6);
-}
-// Central finite difference approximation
-template<typename T>
-T numerical_derivative(T (*f)(T), T x) {
-    const T h = get_h<T>();
-    return (f(x + h) - f(x - h)) / (2 * h);
-}
-
-// Test function
-template<typename T>
-void test_func(const std::string &name,
-               T (*f)(T), const clad::CladFunction<T(*)(T)> &clad_fwd,
-               const decltype(clad::gradient((T (*)(T))nullptr)) &clad_rev,
-               const T (&test_points)[7] = {-2.0,-1.0,-0.5,0.0,0.5,1.0,2.0}) {
-    const T tol = get_tolerance<T>();
-    std::cerr << std::setprecision(10);
-    for (T x : test_points) {
-        T dfwd_val = clad_fwd.execute(x);
-        T drev_val = T();
-        clad_rev.execute(x, &drev_val);
-        T df_num = numerical_derivative(f, x);
-        if (dfwd_val != drev_val || std::abs(dfwd_val - df_num) > tol) {
-            std::cerr << "FAIL: " << name << " at x=" << x
-                      << ": dfwd=" << dfwd_val
-                      << ", drev=" << drev_val
-                      << ", numerical=" << df_num << "\n";
-        } else {
-            std::cout << "PASS: " << name << " at x=" << x
-                      << " dfwd=" << dfwd_val << ", drev=" << drev_val << "\n";
-        }
-    }
-}
+#include "../TestUtils.h"
 
 #define STRINGIFY2(X) #X
 #define STRINGIFY(X) STRINGIFY2(X)
-#define CHECK_F(X) test_func<float>(STRINGIFY(X), \
+#define CHECK_F(X) test_utils::test_func<float>(STRINGIFY(X), \
                            X, clad::differentiate(X), \
                            clad::gradient(X));
-#define CHECK_D(X) test_func<double>(STRINGIFY(X), \
+#define CHECK_D(X) test_utils::test_func<double>(STRINGIFY(X), \
                            X, clad::differentiate(X), \
                            clad::gradient(X));
-#define CHECK_LD(X) test_func<long double>(STRINGIFY(X), \
+#define CHECK_LD(X) test_utils::test_func<long double>(STRINGIFY(X), \
                            X, clad::differentiate(X), \
                            clad::gradient(X));
 #define CHECK_RANGE_F(X, ...)                           \
   {                                                     \
     float arr[] = __VA_ARGS__;                          \
-    test_func<float>(STRINGIFY(X),                      \
+    test_utils::test_func<float>(STRINGIFY(X),                      \
                      X, clad::differentiate(X),         \
                      clad::gradient(X), arr);           \
   }
 #define CHECK_RANGE_D(X, ...)                           \
   {                                                     \
     double arr[] = __VA_ARGS__;                         \
-    test_func<double>(STRINGIFY(X),                     \
+    test_utils::test_func<double>(STRINGIFY(X),                     \
                       X, clad::differentiate(X),        \
                       clad::gradient(X), arr);          \
   }
 #define CHECK_RANGE_LD(X, ...)                          \
   {                                                     \
     long double arr[] = __VA_ARGS__;                    \
-    test_func<long double>(STRINGIFY(X),                \
+    test_utils::test_func<long double>(STRINGIFY(X),                \
                            X, clad::differentiate(X),   \
                            clad::gradient(X), arr);     \
   }
